rework how the samples buffer gets filled
if the received sample rate and channel count match, we can make use of the ndi api AudioFrameToInterleaved() and then memcpy the data.
otherwise we can work with the interleaved data in a direct manner as we need to grab single samples anyway to do resampling or remapping.

for now the remapping is just writing the same mix of all input channels to all output channels.

the resampling uses lerping between two samples. (might need an additional upper bound check.)